docs: add zh_CN translation
