+++ docs/zh_CN/config/index.md
@@ -1,0 +1,15 @@
+# Configurations
+
+Configurations are what make UnoCSS powerful.
+
+- [Rules](/config/rules) - Define atomic CSS utilities
+- [Shortcuts](/config/shortcuts) - Combine multiple rules into a single shorthand.
+- [Theme](/config/theme) - Define theme variables.
+- [Variants](/config/variants) - Apply custom conventions to rules.
+- [Transformers](/config/transformers) - Code transformers to user sources code to support conventions.
+- [Extractors](/config/extractors) - Define where and how the usage of utilities are extracted.
+- [Preflights](/config/preflights) - Define global CSS rules.
+- [Layers](/config/layers) - Define the order of each utilities layer.
+- [Autocomplete](/config/autocomplete) - Define customized autocomplete suggestions.
+- [Presets](/config/presets) - Predefined configurations for common use cases.
+
