+++ docs/zh_CN/config/shortcuts.md
@@ -1,0 +1,64 @@
+---
+title: Shortcuts
+description: The shortcuts functionality that UnoCSS provides is similar to Windi CSS's one.
+---
+
+# Shortcuts
+
+Shortcuts could let you combine multiple rules into a single shorthand, inspired by [Windi CSS's](https://windicss.org/features/shortcuts.html).
+
+## Usage
+
+<!--eslint-skip-->
+
+```ts
+shortcuts: {
+  // shortcuts to multiple utilities
+  'btn': 'py-2 px-4 font-semibold rounded-lg shadow-md',
+  'btn-green': 'text-white bg-green-500 hover:bg-green-700',
+  // single utility alias
+  'red': 'text-red-100'
+}
+```
+
+In addition to the plain mapping, UnoCSS also allows you to define dynamic shortcuts.
+
+Similar to [Rules](/config/rules), a dynamic shortcut is the combination of a matcher RegExp and a handler function.
+
+```ts
+shortcuts: [
+  // you could still have object style
+  {
+    btn: 'py-2 px-4 font-semibold rounded-lg shadow-md',
+  },
+  // dynamic shortcuts
+  [/^btn-(.*)$/, ([, c]) => `bg-${c}-400 text-${c}-100 py-2 px-4 rounded-lg`],
+]
+```
+
+With this, we could use `btn-green` and `btn-red` to generate the following CSS:
+
+```css
+.btn-green {
+  padding-top: 0.5rem;
+  padding-bottom: 0.5rem;
+  padding-left: 1rem;
+  padding-right: 1rem;
+  --un-bg-opacity: 1;
+  background-color: rgba(74, 222, 128, var(--un-bg-opacity));
+  border-radius: 0.5rem;
+  --un-text-opacity: 1;
+  color: rgba(220, 252, 231, var(--un-text-opacity));
+}
+.btn-red {
+  padding-top: 0.5rem;
+  padding-bottom: 0.5rem;
+  padding-left: 1rem;
+  padding-right: 1rem;
+  --un-bg-opacity: 1;
+  background-color: rgba(248, 113, 113, var(--un-bg-opacity));
+  border-radius: 0.5rem;
+  --un-text-opacity: 1;
+  color: rgba(254, 226, 226, var(--un-text-opacity));
+}
+```
