+++ docs/zh_CN/integrations/index.md
@@ -1,0 +1,13 @@
+# Integrations
+
+### Frameworks / Tools
+
+UnoCSS comes with integrations for various frameworks / tools:
+
+<ContentIntegrations />
+
+### Examples
+
+Source code for all the examples can be found in the [/examples](https://github.com/unocss/unocss/tree/main/examples) directory.
+
+<ContentExamples/>
