+++ docs/zh_CN/integrations/jetbrains.md
@@ -1,0 +1,27 @@
+---
+title: Community JetBrains Plugin for UnoCSS
+---
+
+# JetBrains IDEs Plugin
+
+A community [plugin](https://github.com/re-ovo/unocss-intellij) that provides support for Unocss in JetBrains IDEs.
+
+:::info
+Note: This is a **community-maintained plugin, not reviewed or maintained by the UnoCSS team**. Please do your own research before using. If you have any issues with this plugin, please report it to [re-ovo/unocss-intellij](https://github.com/re-ovo/unocss-intellij).
+:::
+
+## Installation
+
+[JetBrains Marketplace](https://plugins.jetbrains.com/plugin/22204-unocss)
+
+## Features
+
+- Auto-completion for UnoCSS classes
+- Hover documentation for UnoCSS classes
+- Code folding
+- Color/Icon preview
+
+## Bug Reports / Feature Requests
+
+Report it at [Issue Tracker](https://github.com/re-ovo/unocss-intellij/issues)
+
