+++ docs/zh_CN/presets/index.md
@@ -1,0 +1,32 @@
+# Official Packages
+
+## Presets
+
+| Package | Description |
+| ------- | ----------- |
+| [@unocss/preset-uno](/presets/uno) | The default preset |
+| [@unocss/preset-mini](/presets/mini) | The minimal but essential rules and variants |
+| [@unocss/preset-wind](/presets/wind) | Tailwind / Windi CSS compact preset |
+| [@unocss/preset-attributify](/presets/attributify) | Enables Attributify Mode for other rules |
+| [@unocss/preset-tagify](/presets/tagify) | Enables Tagify Mode for other rules |
+| [@unocss/preset-icons](/presets/icons) | Pure CSS Icons solution powered by Iconify |
+| [@unocss/preset-web-fonts](/presets/web-fonts) | Web fonts (Google Fonts, etc.) support |
+| [@unocss/preset-typography](/presets/typography) | The typography preset |
+| [@unocss/preset-rem-to-px](/presets/rem-to-px) | Converts rem to px for utils |
+
+## Transformers
+
+| Package | Description |
+| ------- | ----------- |
+| [@unocss/transformer-variant-group](/transformers/variant-group) | Transformer for Windi CSS's variant group feature | ✅ |
+| [@unocss/transformer-directives](/transformers/directives) | Transformer for CSS directives like `@apply` | ✅ |
+| [@unocss/transformer-compile-class](/transformers/compile-class) | Compile group of classes into one class | ✅ |
+| [@unocss/transformer-attributify-jsx](/transformers/attributify-jsx) | Support valueless attributify in JSX/TSX | ✅ |
+
+## Extractors
+
+| Package | Description |
+| ------- | ----------- |
+| [@unocss/extractor-pug](/extractors/pug) | Extractor for Pug |
+| [@unocss/extractor-svelte](/extractors/svelte) | Extractor for Svelte |
+| [@unocss/extractor-arbitrary-variants](/extractors/arbitrary-variants) | Extractor to support arbitrary variants for utilities |
