+++ docs/zh_CN/tools/inspector.md
@@ -1,0 +1,18 @@
+---
+title: Inspector
+description: The inspector UI for UnoCSS (@unocss/inspector).
+---
+
+# Inspector
+
+The inspector UI for UnoCSS: `@unocss/inspector`.
+Ships with `unocss` and `@unocss/vite`.
+
+## Usage
+
+Visit <a href="http://localhost:5173/__unocss" target="_blank" rel="noreferrer">localhost:5173/__unocss</a> in your Vite dev server to see the inspector.
+
+The inspector allows you to inspect the generated CSS rules and the applied classes for each file. And also provided a REPL to test your utilities based on your current configuration.
+
+<img src="https://user-images.githubusercontent.com/11247099/140885990-1827f5ce-f12a-4ed4-9d63-e5145a65fb4a.png" loading="lazy">
+<img src="https://user-images.githubusercontent.com/11247099/140886020-7014f412-f020-4aed-a169-d025cc1bbcd3.png" loading="lazy">
